Time: 47 ms (19.53%) | Memory: 58.4 MB (79.30%) - LeetSync

+++ 1996-number-of-ways-to-rearrange-sticks-with-k-sticks-visible/number-of-ways-to-rearrange-sticks-with-k-sticks-visible.java
@@ -4,7 +4,7 @@ class Solution {
         if(n < k || (k == 0 && n != 0)) return 0;
         if(n == 0 && k== 0) return 1;
         if(dp[n][k] != -1) return dp[n][k];
-        long ans = (sub(n-1 , k-1, dp) + (long)(n-1) * sub(n-1 , k , dp)) % mod; 
+        long ans = ((long)sub(n-1 , k-1, dp) + (long)(n-1) * sub(n-1 , k , dp)) % mod; 
         return  dp[n][k] = (int)ans;
     }
     public int rearrangeSticks(int n, int k) {
